+++ CHANGELOG.md
@@ -107,6 +107,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.40] - 2026-03-30
+
+### Other
+- Update pyqual/builtin_collectors.py
+- Update pyqual/plugins.py
+
 ## [0.1.39] - 2026-03-30
 
 ### Docs